requirements.txt
pandas
mysql-connector-python
sqlalchemy
google-cloud-bigquery
